pantry-app/backend/requirements.txt
flask
flask-cors
python-dotenv
requests
anthropic
openai
Pillow
